Fix asset scripts

+++ assets_server/mappers.py
@@ -84,11 +84,12 @@ class FileManager:
         )
 
     def delete(self, file_path):
-        self.swift_connection.delete_object(
-            self.container_name,
-            normalize(file_path)
-        )
-        return True
+        if self.exists(file_path):
+            self.swift_connection.delete_object(
+                self.container_name,
+                normalize(file_path)
+            )
+            return True
 
     def generate_asset_path(self, file_data, friendly_name):
         """
@@ -169,7 +170,8 @@ class DataManager:
         ]
 
     def delete(self, file_path):
-        self.data_collection.remove({'file_path': file_path})
+        if self.exists(file_path):
+            return self.data_collection.remove({'file_path': file_path})
 
 
 class TokenManager:
